Ignore -WrapBodyInArray when body is omitted
The idea is to better handle the case when someone accidentally leaves
`-WrapBodyInArray` on a GET request or something like that.  One downside
is that it removes a way to pass `[]` as the body, but a) hopefully the
Cloud Office API never expects that, and b) I'm not sure that this case
worked before this change.